Extract common Account elements into concern
