+++ spec/models/brand_account_spec.rb
@@ -1,14 +1,14 @@
 # frozen_string_literal: true
 
+require './spec/support/concerns/account_examples.rb'
+
 RSpec.describe BrandAccount, type: :model do
   describe 'Validations' do
     subject(:account) { FactoryBot.create(:brand_account) }
 
-    it { is_expected.to validate_presence_of(:external_uid) }
-    it { is_expected.to validate_presence_of(:email).allow_nil }
     it { is_expected.to validate_presence_of(:provider) }
-    it { is_expected.to validate_uniqueness_of(:external_uid).scoped_to(:provider) }
     it { is_expected.to validate_uniqueness_of(:provider).scoped_to(:brand_id).ignoring_case_sensitivity }
+    it { is_expected.to validate_uniqueness_of(:external_uid).scoped_to(:provider).ignoring_case_sensitivity }
   end
 
   it { is_expected.to define_enum_for(:provider).with_values(%i[twitter disqus]) }
@@ -16,6 +16,8 @@ RSpec.describe BrandAccount, type: :model do
   describe 'Relations' do
     it { is_expected.to belong_to(:brand) }
   end
+
+  it_behaves_like 'account'
 
   describe '.from_omniauth' do
     %w[twitter].each do |provider|
@@ -204,20 +206,6 @@ RSpec.describe BrandAccount, type: :model do
           end
         end
       end
-    end
-  end
-
-  describe '#client' do
-    subject(:client) { account.client }
-
-    let(:account) { FactoryBot.build(:brand_account) }
-
-    context 'when provider is twitter' do
-      before do
-        account.provider = 'twitter'
-      end
-
-      it { is_expected.to be_an_instance_of(Clients::Twitter) }
     end
   end
